Move address dedupe helpers to utils
dedupeAddresses and its helpers sit better alongside
sortAddressesByStateAndCity in utils/utils.js than in extract.js.
